Arreglando detalles al eliminar con fetch

+++ js/crud_fetch.js
@@ -93,22 +93,26 @@ d.addEventListener("click", async e=> {
         $form.id.value = e.target.dataset.id
     }
     if(e.target.matches(".delete")){
-        //Delete - DELETE
-        try{
-            let options = {
-                method:"DELETE",
-                headers: {
-                    "Content-type":"application/json; charset=utf-8"
-                }
-            },
-            res= await fetch(`http://localhost:5555/santos/${e.target.dataset.id}`, options),
-            json = await res.json();
 
-            if(!res.ok)throw {status: res.status, statusText: res.statusText}
-            location.reload();
-        }catch(err){
-            let message = err.statusText || "Ocurrió un error"
-            $form.insertAdjacentHTML("afterend", `<p><b>Error ${err.status}: ${message}</b></p>`)
+        let isDelete = confirm(`¿EStás seguro de eliminar el id ${e.target.dataset.id}?`);
+        //Delete - DELETE
+        if(isDelete){
+            try{
+                let options = {
+                    method:"DELETE",
+                    headers: {
+                        "Content-type":"application/json; charset=utf-8"
+                    }
+                },
+                res= await fetch(`http://localhost:5555/santos/${e.target.dataset.id}`, options),
+                json = await res.json();
+
+                if(!res.ok)throw {status: res.status, statusText: res.statusText}
+                location.reload();
+            }catch(err){
+                let message = err.statusText || "Ocurrió un error"
+                $form.insertAdjacentHTML("afterend", `<p><b>Error ${err.status}: ${message}</b></p>`)
+            }
         }
     }
 })
